+++ E-Commerce.Application/Features/Order/Handlers/Commands/DeleteOrderCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using E_Commerce.Application.Exceptions;
 using E_Commerce.Application.Features.Order.Requests.Commands;
-using E_Commerce.Application.Persistence.Contracts;
+using E_Commerce.Application.Contracts.Persistence;
+using E_Commerce.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -22,6 +24,12 @@ namespace E_Commerce.Application.Features.Order.Handlers.Commands
         public async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
         {
             var order = await _orderRepository.Get(request.Id);
+
+            if (order == null)
+            {
+                throw new NotFoundException(nameof(Domain.Order), request.Id);
+            }
+
             await _orderRepository.Delete(order);
             return Unit.Value;
         }
