+++ E-Commerce.Application/Features/OrderItem/Handlers/Commands/DeleteOrderItemCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
+using E_Commerce.Application.Exceptions;
 using E_Commerce.Application.Features.OrderItem.Requests.Commands;
-using E_Commerce.Application.Persistence.Contracts;
+using E_Commerce.Application.Contracts.Persistence;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -22,6 +23,12 @@ namespace E_Commerce.Application.Features.OrderItem.Handlers.Commands
         public async Task<Unit> Handle(DeleteOrderItemCommand request, CancellationToken cancellationToken)
         {
             var orderItem = await _orderItemRepository.Get(request.Id);
+
+            if(orderItem == null)
+            {
+                throw new NotFoundException(nameof(Domain.OrderItem), request.Id);
+            }
+
             await _orderItemRepository.Delete(orderItem);
             return Unit.Value;
         }
