+++ E-Commerce.Application/Features/Transaction/Handlers/Commands/DeleteTransactionCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using E_Commerce.Application.Exceptions;
 using E_Commerce.Application.Features.Transaction.Requests.Commands;
-using E_Commerce.Application.Persistence.Contracts;
+using E_Commerce.Application.Contracts.Persistence;
+using E_Commerce.Domain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -24,6 +26,10 @@ namespace E_Commerce.Application.Features.Transaction.Handlers.Commands
         public async Task<Unit> Handle(DeleteTransactionCommand request, CancellationToken cancellationToken)
         {
             var transaction = await _transactionRepository.Get(request.Id);
+
+            if (transaction == null)
+                throw new NotFoundException(nameof(Domain.Transaction), request.Id);
+
             await _transactionRepository.Delete(transaction);
             return Unit.Value;
         }
